refactor: add AppWrapper
