+++ src/components/AppWrapper/AppWrapper.module.scss
@@ -1,0 +1,3 @@
+.appWrapper {
+  
+}
